requirements.txt
# Core LangChain packages
langchain-aws>=0.2.0
langchain-core>=0.3.0
langchain-chroma>=0.1.0

# Vector Database
chromadb>=1.3.0

# AWS Services
boto3>=1.42.0

# PDF Processing
pdfplumber>=0.11.0

# Environment & Configuration
python-dotenv>=1.0.0

# FastAPI Backend
fastapi>=0.124.0
uvicorn[standard]>=0.38.0
python-multipart>=0.0.20
pydantic>=2.0.0

# Streamlit Frontend
streamlit>=1.50.0
requests>=2.32.0
